fix: renaming and refactoring as well as licencing files
